move: 컴포넌트 기본이 되는 클래스 base 디렉토리 밑으로 이동
