+++ views/javascript/components/base/basePopup.js
@@ -1,4 +1,4 @@
-import { DomManager } from '../utils/domManager.js';
+import { DomManager } from '../../utils/domManager.js';
 import { BaseComponent } from './baseComponent.js';
 
 export class BasePopup extends BaseComponent {
