--- views/javascript/components/baseSlide.js
@@ -1,11 +1,0 @@
-import { BaseComponent } from './baseComponent.js';
-
-export class BaseSlide extends BaseComponent {
-  constructor(holder) {
-    super(holder);
-  }
-
-  toggle() {
-    this.holder.classList.toggle('active');
-  }
-}
